refactor folder name and lat lon enum
