--- dl4j-labs/src/main/java/global/skymind/solution/regression/demandRegression/CoordinatesType.java
@@ -1,26 +1,0 @@
-package global.skymind.solution.regression.demandRegression;/*
- *
- *  * ******************************************************************************
- *  *  * Copyright (c) 2019 Skymind AI Bhd.
- *  *  * Copyright (c) 2020 CertifAI Sdn. Bhd.
- *  *  *
- *  *  * This program and the accompanying materials are made available under the
- *  *  * terms of the Apache License, Version 2.0 which is available at
- *  *  * https://www.apache.org/licenses/LICENSE-2.0.
- *  *  *
- *  *  * Unless required by applicable law or agreed to in writing, software
- *  *  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- *  *  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- *  *  * License for the specific language governing permissions and limitations
- *  *  * under the License.
- *  *  *
- *  *  * SPDX-License-Identifier: Apache-2.0
- *  *  *****************************************************************************
- *
- *
- */
-
-
-public enum CoordinatesType {
-    LAT, LON
-}
