+++ dl4j-labs/src/main/java/global/skymind/solution/regression/demandregression/RidershipDemandRegression.java
@@ -1,4 +1,4 @@
-package global.skymind.solution.regression.demandRegression;/*
+package global.skymind.solution.regression.demandregression;/*
  *
  *  * ******************************************************************************
  *  *  * Copyright (c) 2019 Skymind AI Bhd.
@@ -47,7 +47,6 @@ import org.nd4j.linalg.dataset.api.iterator.DataSetIterator;
 import org.nd4j.linalg.dataset.api.preprocessor.DataNormalization;
 import org.nd4j.linalg.dataset.api.preprocessor.NormalizerStandardize;
 import org.nd4j.linalg.learning.config.Adam;
-import org.nd4j.linalg.learning.config.Nesterovs;
 import org.nd4j.linalg.lossfunctions.LossFunctions;
 import org.nd4j.util.ArchiveUtils;
 import org.slf4j.Logger;
@@ -61,7 +60,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.regex.Pattern;
 
 // Features to try
 // (latlong.getLat() + 90)*180 + latlong.getLon()
